pyproject.toml
[tool.poetry]
name = "pyintergraph"
version = "1.3.2"
authors = ["Lukas Erhard <luerhard@googlemail.com>"]
repository = "https://github.com/luerhard/pyintergraph"
homepage = "https://github.com/luerhard/pyintergraph"
description = "Convert Python-Graph-Objects between networkx, python-igraph and graph-tools."
readme="README.md"
license="MIT"
classifiers = [
        "Topic :: Utilities",
        "Intended Audience :: Science/Research",
    ]
packages = [{include = "pyintergraph"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
numpy = [{version = ">=1.26", python = ">=3.12"}, {version = ">=1.18", python = "<3.12"}]
networkx = {version = ">=2.4", optional = true}
python-igraph = {version = ">=0.8", optional = true}

[tool.poetry.extras]
networkx = ["networkx"]
python-igraph = ["python-igraph"]
net = ["networkx", "python-igraph"]

[tool.poetry.group.dev.dependencies]
pytest = ">=7"
pytest-cov = ">=4"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length=100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
